+++ src/DebugCodeLens/index.ts
@@ -1,1 +1,2 @@
 export { DebugCodeLensProvider } from './DebugCodeLensProvider'
+export { TestState } from './TestState'