+++ tests/JestExt.test.ts
@@ -5,13 +5,12 @@ jest.mock('../src/DebugCodeLens', () => ({
   DebugCodeLensProvider: class MockCodeLensProvider {},
 }))
 
-import { EventEmitter } from 'events'
 import { JestExt } from '../src/JestExt'
 import { ProjectWorkspace, Settings } from 'jest-editor-support'
 import { window, workspace, debug } from 'vscode'
 import { hasDocument, isOpenInMultipleEditors } from '../src/editor'
 import { failingAssertionStyle } from '../src/decorations'
-import { resultsWithLowerCaseWindowsDriveLetters } from '../src/TestResults'
+import { POINT_CONVERSION_COMPRESSED } from 'constants'
 
 describe('JestExt', () => {
   const mockSettings = (Settings as any) as jest.Mock<any>
@@ -19,7 +18,7 @@ describe('JestExt', () => {
   let projectWorkspace: ProjectWorkspace
   const channelStub = { appendLine: () => {} } as any
   const mockShowErrorMessage = window.showErrorMessage as jest.Mock<any>
-  const extensionSettings = {} as any
+  const extensionSettings = { debugCodeLens: {} } as any
 
   beforeEach(() => {
     jest.resetAllMocks()
@@ -33,7 +32,7 @@ describe('JestExt', () => {
       getConfig: callback => callback(),
       jestVersionMajor: 17,
     }))
-    new JestExt(projectWorkspace, channelStub, extensionSettings)
+    new JestExt(null, projectWorkspace, channelStub, extensionSettings)
 
     expect(mockShowErrorMessage.mock.calls).toMatchSnapshot()
   })
@@ -43,7 +42,7 @@ describe('JestExt', () => {
       getConfig: callback => callback(),
       jestVersionMajor: 20,
     }))
-    new JestExt(projectWorkspace, channelStub, extensionSettings)
+    new JestExt(null, projectWorkspace, channelStub, extensionSettings)
     expect(window.showErrorMessage).not.toBeCalled()
   })
 
@@ -56,7 +55,7 @@ describe('JestExt', () => {
     const decorationType: any = { dispose: jest.fn() }
 
     beforeEach(() => {
-      sut = new JestExt(projectWorkspace, channelStub, extensionSettings)
+      sut = new JestExt(null, projectWorkspace, channelStub, extensionSettings)
 
       sut.canUpdateDecorators = jest.fn().mockReturnValueOnce(true)
       sut.debugCodeLensProvider.didChange = jest.fn()
@@ -105,8 +104,11 @@ describe('JestExt', () => {
 
   describe('generateInlineErrorDecorator()', () => {
     it('should add the decoration type to the cache', () => {
-      const settings: any = { enableInlineErrorMessages: true }
-      const sut = new JestExt(projectWorkspace, channelStub, settings)
+      const settings: any = {
+        debugCodeLens: {},
+        enableInlineErrorMessages: true,
+      }
+      const sut = new JestExt(null, projectWorkspace, channelStub, settings)
       const editor: any = {
         document: { fileName: 'file.js' },
         setDecorations: jest.fn(),
@@ -134,31 +136,37 @@ describe('JestExt', () => {
   describe('runTest()', () => {
     const fileName = 'fileName'
     const testNamePattern = 'testNamePattern'
-    const defaultArgs = ['--runInBand', fileName, '--testNamePattern', testNamePattern]
 
-    it('should use the config if set', () => {
-      const config = 'jest.json'
-      const expected = [...defaultArgs, '--config', config]
-      const extensionSettings = {
-        pathToConfig: config,
-      } as any
+    it('should run the supplied test', async () => {
+      const startDebugging = debug.startDebugging as jest.Mock<Function>
 
-      const sut = new JestExt(projectWorkspace, channelStub, extensionSettings)
-      // @ts-ignore: Overriding private method
-      sut.resolvePathToJestBin = jest.fn().mockReturnValueOnce(true)
-      sut.runTest(fileName, testNamePattern)
+      startDebugging.mockImplementation(async (_folder: any, nameOrConfig: any) => {
+        // trigger fallback to default configuration
+        if (typeof nameOrConfig === 'string') {
+          throw null
+        }
+      })
 
-      expect(debug.startDebugging).toHaveBeenCalledTimes(1)
+      const sut = new JestExt(null, projectWorkspace, channelStub, extensionSettings)
+      ;(sut.debugConfigurationProvider.provideDebugConfigurations as jest.Mock<Function>).mockReturnValue([
+        { type: 'dummyconfig' },
+      ])
 
-      const configuration = (debug.startDebugging as jest.Mock<Function>).mock.calls[0][1]
+      await sut.runTest(fileName, testNamePattern)
+
+      expect(debug.startDebugging).toHaveBeenCalled()
+
+      const configuration = startDebugging.mock.calls[startDebugging.mock.calls.length - 1][1]
       expect(configuration).toBeDefined()
-      expect(configuration.args).toEqual(expected)
+      expect(configuration.type).toBe('dummyconfig')
+
+      expect(sut.debugConfigurationProvider.prepareTestRun).toBeCalledWith(fileName, testNamePattern)
     })
   })
 
   describe('onDidCloseTextDocument()', () => {
     const projectWorkspace = new ProjectWorkspace(null, null, null, null)
-    const sut = new JestExt(projectWorkspace, channelStub, extensionSettings)
+    const sut = new JestExt(null, projectWorkspace, channelStub, extensionSettings)
     const document = {} as any
     sut.removeCachedTestResults = jest.fn()
     sut.removeCachedDecorationTypes = jest.fn()
@@ -176,7 +184,7 @@ describe('JestExt', () => {
 
   describe('removeCachedTestResults()', () => {
     const projectWorkspace = new ProjectWorkspace(null, null, null, null)
-    const sut = new JestExt(projectWorkspace, channelStub, extensionSettings)
+    const sut = new JestExt(null, projectWorkspace, channelStub, extensionSettings)
     sut.testResultProvider.removeCachedResults = jest.fn()
 
     it('should do nothing when the document is falsy', () => {
@@ -201,7 +209,7 @@ describe('JestExt', () => {
 
   describe('removeCachedAnnotations()', () => {
     const projectWorkspace = new ProjectWorkspace(null, null, null, null)
-    const sut = new JestExt(projectWorkspace, channelStub, extensionSettings)
+    const sut = new JestExt(null, projectWorkspace, channelStub, extensionSettings)
 
     beforeEach(() => {
       sut.failingAssertionDecorators = {
@@ -227,7 +235,7 @@ describe('JestExt', () => {
     let sut
     const editor: any = {}
     const projectWorkspace = new ProjectWorkspace(null, null, null, null)
-    sut = new JestExt(projectWorkspace, channelStub, extensionSettings)
+    sut = new JestExt(null, projectWorkspace, channelStub, extensionSettings)
     sut.triggerUpdateDecorations = jest.fn()
 
     beforeEach(() => {
@@ -261,7 +269,7 @@ describe('JestExt', () => {
 
     beforeEach(() => {
       const projectWorkspace = new ProjectWorkspace(null, null, null, null)
-      sut = new JestExt(projectWorkspace, channelStub, extensionSettings)
+      sut = new JestExt(null, projectWorkspace, channelStub, extensionSettings)
     })
 
     function expectItTakesNoAction(event) {
@@ -328,7 +336,7 @@ describe('JestExt', () => {
 
   describe('toggleCoverageOverlay()', () => {
     it('should toggle the coverage overlay visibility', () => {
-      const sut = new JestExt(projectWorkspace, channelStub, extensionSettings)
+      const sut = new JestExt(null, projectWorkspace, channelStub, extensionSettings)
       sut.toggleCoverageOverlay()
 
       expect(sut.coverageOverlay.toggleVisibility).toBeCalled()
@@ -339,7 +347,7 @@ describe('JestExt', () => {
     it('should update the coverage overlay in visible editors', () => {
       const editor: any = {}
 
-      const sut = new JestExt(projectWorkspace, channelStub, extensionSettings)
+      const sut = new JestExt(null, projectWorkspace, channelStub, extensionSettings)
       sut.triggerUpdateDecorations(editor)
 
       expect(sut.coverageOverlay.updateVisibleEditors).toBeCalled()