+++ tests/extension.test.ts
@@ -4,7 +4,9 @@ jest.mock('vscode', () => ({
   CodeLens: class {},
   commands: {
     registerCommand: jest.fn(),
-    registerTextEditorCommand: jest.fn(),
+  },
+  debug: {
+    registerDebugConfigurationProvider: jest.fn(),
   },
   languages: {
     registerCodeLensProvider: jest.fn(),
@@ -18,8 +20,21 @@ jest.mock('vscode', () => ({
     onDidChangeActiveTextEditor: jest.fn().mockReturnValue('onDidChangeActiveTextEditor'),
   },
   workspace: {
-    getConfiguration: jest.fn().mockReturnValue({
-      get: jest.fn().mockImplementation(key => key),
+    /** Mock getConfiguration by reading default values from package.json */
+    getConfiguration: jest.fn().mockImplementation(section => {
+      const data = readFileSync('./package.json')
+      const config = JSON.parse(data.toString()).contributes.configuration.properties
+
+      const defaults = {}
+      for (const key of Object.keys(config)) {
+        if (section.length === 0 || key.startsWith(`${section}.`)) {
+          defaults[key] = config[key].default
+        }
+      }
+
+      return {
+        get: jest.fn().mockImplementation(key => defaults[`${section}.${key}`]),
+      }
     }),
     onDidChangeConfiguration: jest.fn(),
     onDidCloseTextDocument: jest.fn(),
@@ -47,8 +62,10 @@ jest.mock('../src/SnapshotCodeLens', () => ({
   registerSnapshotPreview: jest.fn(() => []),
 }))
 
-import { activate } from '../src/extension'
+import { activate, getExtensionSettings } from '../src/extension'
 import * as vscode from 'vscode'
+import { readFileSync } from 'fs'
+import { TestState } from '../src/DebugCodeLens'
 
 describe('Extension', () => {
   describe('activate()', () => {
@@ -88,6 +105,42 @@ describe('Extension', () => {
 
       expect((vscode.commands.registerCommand as jest.Mock<any>).mock.calls).toContainEqual(expected)
       expect(context.subscriptions.push.mock.calls[0]).toContain(expected)
+    })
+
+    it('should register a DebugConfigurationProvider', () => {
+      const register = vscode.debug.registerDebugConfigurationProvider as jest.Mock<any>
+      register.mockReset()
+
+      activate(context)
+
+      expect(register).toHaveBeenCalledTimes(2)
+      const registeredAsNode = register.mock.calls.some(parameters => parameters[0] === 'node')
+      const registeredAsJestTest = register.mock.calls.some(parameters => parameters[0] === 'vscode-jest-tests')
+      expect(registeredAsNode && registeredAsJestTest).toBeTruthy()
+    })
+  })
+
+  describe('getExtensionSettings()', () => {
+    it('should return the extension configuration', async () => {
+      vscode.workspace.rootPath = '<rootDir>'
+
+      expect(getExtensionSettings()).toEqual({
+        autoEnable: true,
+        coverageFormatter: 'DefaultFormatter',
+        debugCodeLens: {
+          enabled: true,
+          showWhenTestStateIn: [TestState.Fail, TestState.Unknown],
+        },
+        enableInlineErrorMessages: true,
+        enableSnapshotPreviews: true,
+        enableSnapshotUpdateMessages: true,
+        pathToConfig: '',
+        pathToJest: null,
+        restartJestOnSnapshotUpdate: false,
+        rootPath: '<rootDir>',
+        runAllTestsFirst: true,
+        showCoverageOnLoad: false,
+      })
     })
   })
 })